+++ Runtime/iOS/CASMediationManager.cs
@@ -46,7 +46,7 @@ namespace CAS.iOS
         public CASMediationManager( CASInitSettings initData )
         {
             managerID = initData.targetId;
-            isTestAdMode = initData.testAdMode;
+            isTestAdMode = initData.IsTestAdMode();
         }
 
         ~CASMediationManager()